% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{getResult}
\alias{getResult}
\alias{getResult,ANY-method}
\title{Gets the result of integral approximation}
\usage{
getResult(object)
}
\arguments{
\item{object}{An object of class Simpson or Trapezoid}
}
\value{
An object of class Trapezoid or Simpson containing
\item{result}{the approximated result}
}
\description{
Gets the result from the Trapezoid and Simpson Rules' calculation of the integral
}
\note{
This method is used by the integrateIt and print methods
}
\seealso{
\code{\link{trapezoid_fun}}, \code{\link{simpson_fun}}, \code{\link{integrateIt}}, \code{\link{print}}
}
\author{
Kimberly Acquilano
}
