% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getX.R
\name{getX}
\alias{getX}
\alias{getX,ANY-method}
\title{Gets the vector of values under the integral}
\usage{
getX(object)
}
\arguments{
\item{object}{An object of class Simpson or Trapezoid}
}
\value{
An object of class Trapezoid or Simpson containing
\item{x}{the vector of values}
}
\description{
Gets the vector of values under the curve for the Trapezoid and Simpson Rules' calculation of the integral
}
\note{
This method is used by the integrateIt and print methods
}
\seealso{
\code{\link{trapezoid_fun}}, \code{\link{simpson_fun}}, \code{\link{integrateIt}}, \code{\link{print}}
}
\author{
Kimberly Acquilano
}
