% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getY.R
\name{getY}
\alias{getY}
\alias{getY,ANY-method}
\title{Gets the vector of evaluated values under the integral}
\usage{
getY(object)
}
\arguments{
\item{object}{An object of class Simpson or Trapezoid}
}
\value{
An object of class Trapezoid or Simpson containing
\item{y}{the vector of evaluated values}
}
\description{
Gets the vector of evaluated values under the curve for the Trapezoid and Simpson Rules' calculation of the integral
}
\note{
This method is used by the integrateIt and print methods
}
\seealso{
\code{\link{trapezoid_fun}}, \code{\link{simpson_fun}}, \code{\link{integrateIt}}, \code{\link{print}}
}
\author{
Kimberly Acquilano
}
