% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_it_classes.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{Trapezoid/Simpson-class,}
\alias{Trapezoid/Simpson-method,}
\alias{integrateIt/print}
\title{Integral approximation classes}
\description{
Class of type "Trapezoid" or of type "Simpson"
}
\details{
Integral objects have the following slots:
\itemize{
\item \code{n} The number of segments of integration
\item \code{f} The function being integrated
\item \code{ab} A vector containing the start and end point of integration
\item \code{rule} A specification of which integration estimation rule to use
}
}
\author{
Jordon Newton
}
