% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate function.R
\name{integrateIt}
\alias{integrateIt}
\alias{addSquares,ANY-method}
\title{Integrates}
\usage{
integrateIt(x, y, start_end_values, test)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{start_end_values}{A numeric}

\item{test}{a character}
}
\value{
An object of class Trap or Simpson
}
\description{
Integrates
}
\note{
Please Work
}
\examples{

}
\seealso{
\code{\link{subtractSquares}}
}
\author{
Jack Ploshnick
}
