\name{integrateIt-package}
\alias{integrateIt-package}
\alias{integrateIt}
\docType{package}
\title{
\packageTitle{integrateIt}
}
\description{
\packageDescription{integrateIt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{integrateIt}
\packageIndices{integrateIt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{integrateIt}

Maintainer: \packageMaintainer{integrateIt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
