% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Trapazoidal or Simpsons integration}
\usage{
integrateIt(x, y, startAndEnd, rule)
}
\arguments{
\item{x}{The vector of x values}

\item{y}{The vector of corresponding y values with the same dimensionality as \code{x}.}

\item{startAndEnd}{The length = 2 vector of the start and end "x" values of the itegration}

\item{rule}{A string value, either "Trap" or "Simpsons" to indicate which integration should be performed}
}
\value{
An object of class Trapazoid or Simpsons, the x and y vectors and the result of the integration
 \item{object}{The original object of class Trapazoid or Simpson}
 \item{XandY}{A factor containing the x and y vectors}
 \item{result}{The result of the integration}
}
\description{
Integrates two vectors of "x" and "y" values via the Trapazoid or Simpsons method
}
\note{
This is a very complicated function function
}
\examples{

x <- c(1,2,3,4,5) 
y <- c(1,4,9,16,25)
startAndEnd <- c(1,5)
rule <- "Trap"
integrateIt(x, y, startAndEnd, rule)
}
\author{
Aaron J. Brezel
}
