% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimpson.R
\docType{methods}
\name{plot,Simpson-method}
\alias{plot,Simpson-method}
\title{Plot objects of class Simpson}
\usage{
\S4method{plot}{Simpson}(x, y, ...)
}
\arguments{
\item{x}{object of class Simpston}
}
\value{
A graph of Simpson's estimation of the integral that was taken between a and b
}
\description{
Inputing an object of class Simpson plots a graph of the integral that was taken
}
\examples{

 s = integrateIt(vector_x = 1:11,
                 vector_y = 1:11,
                 a = 3, b = 7, rule = "Simp")
 plot(s)
}
\author{
David Flaserstein
}

