% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Trapezoid-method}
\alias{print,Trapezoid-method}
\alias{print}
\title{Print function}
\usage{
\S4method{print}{Trapezoid}(x)
}
\arguments{
\item{x}{Object of class Trapezoid or Simpson}
}
\value{
Area under the curve
}
\description{
Print function
}
\note{
Print area under the curve of a calculated integral
}
\seealso{
\code{\link{Trapezoid}}

\code{\link{Simpson}}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
