% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_It.R
\docType{methods}
\name{print,ANY-method}
\alias{print,ANY-method}
\title{Printing the Simpsons or Trapezoid Integration in the integrateIt function}
\usage{
\S4method{print}{ANY}(object)
}
\arguments{
\item{x}{The object of class "Trapezoid" or "Simpsons" at the returned result from the integrateIt function}
}
\description{
This function prints out the result of integrateIt
}
\examples{

result <- integrateIt(x = c(1,2,3), y = c(1,4,9), from = c(1,3), Rule = "Trapezoid")
print(result)


}
\seealso{
\code{\link{integrateIt}}
}
\author{
Alex Petri
}
