% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\docType{methods}
\name{print,Trapezoid-method}
\alias{print,Trapezoid-method}
\alias{integrateIt}
\title{Prints an integral approximation}
\usage{
\S4method{print}{Trapezoid}(x)
}
\arguments{
\item{x}{An object of class Trapezoid or Simpson}
}
\value{
A numeric
}
\description{
Prints an integral approximation
}
\note{
Should return only the integral approximation
}
\examples{

}
\seealso{
\code{\link{integrateIt}}
}
\author{
Kalen Davison
}
