% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIntegrated.R
\docType{methods}
\name{print,integrated-method}
\alias{print,integrated-method}
\title{Prints Trapezoid or Simpson objects}
\usage{
\S4method{print}{integrated}(x, ...)
}
\arguments{
\item{x}{object of class Trapezoid, Simpson, or integreated}
}
\value{
The integrated_value of the Object
}
\description{
Inputting an object of class Trapezoid or Simpson prints out the integrated_value of those objects
}
\examples{

s = integrateIt(vector_x = 1:11,
vector_y = 1:11,
a = 3, b = 7, rule = "Simp")
print(s)
t = integrateIt(vector_x = 1:11,
              vector_y = 1:11,
              a = 3, b = 7, rule = "Trap")
print(t)
}
\author{
David Flaserstein
}

