% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTrap.R
\name{printTrap}
\alias{printTrap}
\alias{printTrap,ANY-method}
\title{Printing result of the Trapezoid Rule}
\usage{
printTrap(object)
}
\arguments{
\item{object}{An object of class Simpson}
}
\value{
An evaluation of the Trapezoid rule for the given object of class Trapezoid
}
\description{
Prints the result of the Trapezoid rule, given an object of class Trapezoid
}
\note{
This is a very simple function
}
\examples{

b<-new("Trapezoid", x=c(1,2,3,4,5), y=c(1,2,3,1,1), ab=c(1,5))
printTrap(b)

}
\seealso{
\code{\link{integrateIt}}
}
\author{
Ian Davis
}
