% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\title{Simpson S4 Class}
\value{

}
\description{
Simpson S4 Class
}
\section{Slots}{

\describe{
\item{\code{equation}}{function.}

\item{\code{x}}{numeric.}

\item{\code{y}}{numeric.}

\item{\code{x1}}{numeric.}

\item{\code{y1}}{numeric.}

\item{\code{n}}{numeric.}

\item{\code{intg}}{numeric.}
}}

