% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson-class.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{Simpson}
\title{An object with two vectors of Class Simpson}
\description{
Object of class \code{Simpson}
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{result} The result of the integral
\item \code{x} a vector of values
\item \code{y} a vector of values of same dimensionality as \code{x} 
}
}
\author{
Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
