% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_fun.R
\name{simpson_fun}
\alias{simpson_fun}
\title{Numerical Integration Using the Simpson Rule}
\usage{
simpson_fun(a, b, d, f)
}
\arguments{
\item{a}{The start value}

\item{b}{The end value}

\item{d}{The number of divisions}

\item{f}{The function that will be used for integration}
}
\value{
A numerical value
}
\description{
Numerical Integration Using the Simpson Rule
}
\examples{
f <- function(x) {sin(x)}
simpson_fun(a = 0, b = 10, d = 5, f)
}
\seealso{
trapezoid_fun
}
