% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsonClass.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{integrateIt}
\title{Creates a Simpson class}
\arguments{
\item{xvalues}{A numeric object}

\item{yvalues}{A numeric object with the same dimensionality as \code{xvalues}.}

\item{area}{A numeric}
}
\value{
creates Simpson Class
}
\description{
Creates a Simpson class
}
\note{
Creates Simpson class with three slots
}
\examples{

}
\seealso{
\code{\link{integrateIt}}
}
\author{
Kalen Davison
}
