% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson function.R
\name{simpsons}
\alias{simpsons}
\alias{addSquares,ANY-method}
\title{Does simpson approximation}
\usage{
simpsons(x, y, StartToEnd)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{StartToEnd}{A numeric}
}
\value{
An object of class Trap or Simpson
}
\description{
Does simpson approximation
}
\note{
Please Work
}
\examples{

}
\seealso{
\code{\link{subtractSquares}}
}
\author{
Jack Ploshnick
}
