% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpsons.R
\docType{class}
\name{Simpsons-class}
\alias{Simpsons-class}
\title{A Simpson's Rule Integration Object}
\description{
Object of class \code{Simpsons} are created by the \code{integrateIt} and \code{print_It} functions
}
\details{
An object of the class "Simpsons" has the following slots:
\itemize{
\item \code{x} A vector of the x values of the input
\item \code{y} A vector of the y values of the input
\item \code{integral} The result of the function
}
}
\author{
Alex Petri: \email{alexpetri@wustl.edu}
}
