% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tol.R
\docType{class}
\name{Tol-class}
\alias{Tol-class}
\alias{Tol-class,}
\alias{initialize,Tol-class,}
\title{A Tol Object}
\description{
Objects of class \code{Tol} are created by the \code{tolTest} function
}
\details{
An object of the class 'Tol' has the following slots:
\itemize{
 \item{fun}{A function}
 \item{tolerance}{A level of tolerance}
 \item{a}{A starting value}
 \item{b}{An ending value}
 \item{rule}{A rule argument, whether the Trapezoidal rule or Simpson rule is used}
 \item{start}{An argument for the number of intervals it should start with}
 \item{correct}{The correct answer for the integral}
 \item{final_n}{The final number of intervals}
 \item{abs_error}{The absolute error of the estimate}
 }
}
\author{
Taishi Muraoka: \email{tmuraoka@wustl.edu}
}

