% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Tolerance Test}
\usage{
tolTest(fun, tolerance, rule, start, Correct, ...)
}
\arguments{
\item{fun}{A vector of values}

\item{tolerance}{A vector of values with the same dimensionality as \code{x}}

\item{rule}{A vector with two values that are the bounds of integration}

\item{start}{Either Trap or Simpson}

\item{Correct}{Either Trap or Simpson}
}
\value{
The following output
 \item{all of the inputs}
 \item{final n}
 \item{the absolute error of the estimate}
}
\description{
take in a function and increase the number of intervals n until the answer it provides using the specified approximation is within tolerance of the correct answer.
}
\note{
This is a very simple function
}
\examples{

integrand <- function(x) {1/((x+1)*sqrt(x))}
tolTest(fun=integrand, tolerance=0.5, rule="Simpson", start=3, correct=integrate(integrand, lower=5,upper=7))

}
\seealso{
\code{\link{print}},\code{\link{integrateIt}}
}
\author{
Benjamin Schneider
}
