% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapSimpson.R
\docType{class}
\name{Trap-class}
\alias{Simpson-class,}
\alias{Trap-class}
\alias{Trap-class,}
\alias{initialize,Trap-Simpsonmethod,}
\alias{print,Trap-Simpson-method}
\alias{show,Trap-Simpson-method,}
\title{A Trap and Simpson Object}
\description{
Objects of class \code{Trap} and \code{Simpson} are created by the \code{integrateIt} function
}
\details{
An object of the class 'Trap' or 'Simpson' has the following slots:
\itemize{
 \item{x}{A vector of x}
 \item{y}{A vector of y}
 \item{result}{The result calculated applying the Trapezoidal or Simpson rule}
 }
}
\author{
Taishi Muraoka: \email{tmuraoka@wustl.edu}
}

