% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapAndSimpsonsClass.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{A Trapezoidal or Simpsons integration object}
\description{
Object of class \code{integrateIt} are created by the \code{integrateIt} functions
}
\details{
An object of the class "Trapezoid" or "Simpsons" has the following slots:
\itemize{
\item \code{x} The vector of x values
\item \code{y} The vector of corresponding y values
\item \code{integral} The result of the integral calculation of the x and y vectors 
}
}
\author{
Aaron J. Brezel: \email{aaronbrezel@wustl.edu}
}
