% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trap Class.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{addSquares,ANY-method}
\title{Creates Trap class}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{estimate}{a numeric}
}
\value{
makes Trap class
}
\description{
Creates Trap class
}
\note{
Please Work
}
\examples{

}
\seealso{
\code{\link{subtractSquares}}
}
\author{
Jack Ploshnick
}
