% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{Trapezoid S4 Class}
\value{

}
\description{
Trapezoid S4 Class
}
\section{Slots}{

\describe{
\item{\code{equation}}{function.}

\item{\code{x}}{numeric.}

\item{\code{y}}{numeric.}

\item{\code{x1}}{numeric.}

\item{\code{y1}}{numeric.}

\item{\code{n}}{numeric.}

\item{\code{intg}}{numeric.}
}}

