% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid-class.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{Trapezoid}
\title{An object with two vectors of Class Trapezoid}
\description{
Object of class \code{Trapezoid}
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{result} The result of the integral
\item \code{x} a vector of values
\item \code{y} a vector of values of same dimensionality as \code{x} 
}
}
\author{
Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
