% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_fun.R
\name{trapezoid_fun}
\alias{trapezoid_fun}
\title{Numerical Integration Using the Trapezoid Rule}
\usage{
trapezoid_fun(a, b, d, f)
}
\arguments{
\item{a}{The start value}

\item{b}{The end value}

\item{d}{The number of divisions}

\item{f}{The function that will be used for integration}
}
\value{
A numerical value
}
\description{
Numerical Integration Using the Trapezoid Rule
}
\examples{
f <- function(x) {sin(x)}
trapezoid_fun(a = 0, b = 10, d = 5, f)

}
\seealso{
simpson_fun
}
