% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoidClass.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{integrateIt}
\title{Creates a Trapezoid class}
\arguments{
\item{xvalues}{A numeric object}

\item{yvalues}{A numeric object with the same dimensionality as \code{xvalues}.}

\item{area}{A numeric}
}
\value{
creates Trapezoid Class
}
\description{
Creates a Trapezoid class
}
\note{
Creates Trapezoid class with three slots
}
\examples{

}
\seealso{
\code{\link{integrateIt}}
}
\author{
Kalen Davison
}
