% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trap function.R
\name{trap}
\alias{trap}
\alias{addSquares,ANY-method}
\title{Does Trap approimation}
\usage{
trap(x, y, StartToEnd)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{StartToEnd}{}
}
\value{
object of class trap
}
\description{
Does Trap approimation
}
\note{
Please Work
}
\examples{

}
\seealso{
\code{\link{subtractSquares}}
}
\author{
Jack Ploshnick
}
