% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{IntegrateIt,ANY-method}
\title{Integrating values}
\usage{
integrateIt(fx, n, interval, rule)
}
\arguments{
\item{fx}{A function defined over \link{a,b}.}

\item{n}{The number of desired subintervals over \link{a,b} used to define x values.}

\item{interval}{The lower and upper bounds of the interval \link{a,b}.}

\item{rule}{\code{'Simpson'} or \code{'Trapezoid'} method of numerical integration.}
}
\value{
An object of class \code{'Simpson'} or \code{'Trapezoid'} containing
\item{result}{The approximated numerical integral}
\item{x}{A subset of the original vector of inputs over \link{a,b}}
\item{y}{The evaluated function values over \link{a,b}}
}
\description{
Integrates function over interval using numerical approximation methods.
}
\note{
Approximation accuracy increases with size of n.
}
\examples{

test_fx <- function(x){x^2}
integrateIt(text_fx, 2000, c(1,3), rule="Trapezoid")

}
\author{
Alma Velazquez
}
