% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{integrateIt function}
\usage{
integrateIt(x, f_x, a, b, rule)
}
\arguments{
\item{x}{A vector of input values}

\item{f_x}{function to generate y with x}

\item{a}{integration starting value}

\item{b}{integration ending value}

\item{rule}{pick method of either trapezoidal or simpson}
}
\value{
Return a list including:
 \item{integral}{the approximated integral}
 \item{x}{vector of values}
 \item{y}{vector of evaluated values}
}
\description{
integrates with trapezoidal or simpsons rule
}
\note{
generic function to produce approximate integral object with Simpson or Trapezoid rule
}
\examples{
x <- seq(0, 10)
y <- seq(0,20,2)
integrateIt(x, y, 0, 10, rule="Trapezoid")

}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}
}
\author{
Annie Jarman
}
