% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integration Techniques}
\usage{
integrateIt(fun, x, a, b, rule)
}
\arguments{
\item{fun}{A function defined by the user.}

\item{x}{A vector of numeric values}

\item{a}{lower bound of integration}

\item{b}{upper bound of integration}

\item{rule}{specified Trapezodial or Simpson rule}
}
\description{
Integrate functions using Trapezoidal Rule or Simpson Rule.
}
