% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integrating a set of x and y values using the Trapezoid or Simpsons Rule}
\usage{
integrateIt(x, y, from, Rule)
}
\arguments{
\item{x}{A numeric object of x-axis values on a curve}

\item{y}{A numeric object with the same dimensionality as \code{x} and corresponding to the y-axis values on a curve}

\item{from}{Two numeric values: a - the minimum value of x, and b - the maximum value of x}

\item{Rule}{A character object: "Trap" to indicate desire to use the Trapezoid Rule and "Simpsons" for the Simpsons Rule}
}
\value{
An object of class Trapezoid or class Simpsons containing the class, the x and y vectors, and the result of the integration
 \item{object}{An object of class Trapezoid Simpson}
}
\description{
Integrates via Trapezoid or Simpsons Rule
}
\examples{


integrateIt(x = c(1,2,3), y = c(1,4,9), from = c(1,3), Rule = "Trap")


}
\seealso{
\code{\link{print}}
}
\author{
Alex Petri
}
