% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Integrate values with estimation}
\usage{
integrateIt(x, y, bounds, rule)
}
\arguments{
\item{x}{A vector of values}

\item{y}{A vector of values with the same dimensionality as \code{x}}

\item{bounds}{A vector with two values that are the bounds of integration}

\item{type}{Either Trap or Simpson}
}
\value{
An object of class Squares containing
 \item{result}{the integrated value}
 \item{x}{The input vector of x's} 
 \item{y}{The input vector of y's}
}
\description{
Finds the sum of squared numbers
}
\note{
This is a very simple function
}
\examples{

X <- seq(0, 10) 
Y <- seq(0,20,2)
integrateIt(X, Y, c(0,10), type="Trap")
}
\seealso{
\code{\link{print}}
}
\author{
Benjamin Schneider
}
