% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integration approximation}
\usage{
integrateIt(x, f, limits, rule, ...)
}
\arguments{
\item{x}{numeric. A vector of values}

\item{f}{numeric. A vector of evaluated values}

\item{limits}{numeric. Limits of integrals: a and b values}

\item{rule}{character. Simpson or Trapezoidal rule}

\item{...}{}
}
\value{

}
\description{
Calculate integral by Simpson or Trapezoidal rule
}
\examples{

x = 1:10
f = 1:10
limits = c(1,10)
integrateIt(x, f, limits, "Simpson")
integrateIt(x, f, limits, "Trapezoid")
}
