% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R, R/print.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\alias{print}
\alias{integrateIt,ANY-method}
\title{Integrating values}
\usage{
integrateIt(x, y, bounds, rule)

print(x, y, bounds, rule, ...)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
An object of class Trapezoid or Simpson containing
 \item{bounds}{The integral of the x and y values}
 \item{x}{The first object input} 
 \item{y}{The second object input}
 \item{trap}
 \item{simp}

An object of class Trapezoid or Simpson containing
 \item{bounds}{The integral of the x and y values}
 \item{x}{The first object input} 
 \item{y}{The second object input}
 \item{trap}
 \item{simp}
}
\description{
Finds the Integrated value of vectors

Print the Integrated value of vectors
}
\note{
This is not a very simple function

This is not a very simple function
}
\examples{

myx<-seq(1,11,1)
myy<-seq(11,21,1)
integrateIt(x,y,bounds=c(1,10), rule="trap")

myx<-seq(1,11,1)
myy<-seq(11,21,1)
integrateIt(x,y,bounds=c(1,10), rule="trap")
}
\author{
Dominique M. Lockett

Dominique M. Lockett
}
