% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Integrating values}
\usage{
integrateIt(x, y, bounds, rule)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
An object of class Trapezoid or Simpson containing
 \item{bounds}{The integral of the x and y values}
 \item{x}{The first object input} 
 \item{y}{The second object input}
 \item{trap}
 \item{simp}
}
\description{
Finds the Integrated value of vectors
}
\note{
This is not a very simple function
}
\examples{

x<-seq(1,11,1)
y<-seq(11,21,1)
bounds<-c(1,10)
rule<-"Trap"
integrateIt(x, y, bounds, rule)
integrateIt(x,y,bounds=c(1,10), rule="trap")
}
\author{
Dominique M. Lockett
}
