% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Integrating}
\usage{
integrateIt(x, y, domain, rule, ...)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{y}{A vector of numeric values, with same dimensionality as \code{x}.}

\item{domain}{A vector of size 2, indicating starting and ending values.}

\item{rule}{Either "Trap" or "Simpson", indicating the integration method to be used.}
}
\value{
An object of class Trapezoid or Simpson containing:
 \item{x}{a vector of x values}
 \item{y}{a vector of y values}
 \item{area}{area under curve defined by x and y, calculated using the indicated method}
}
\description{
Finds the area under a curve given 2 vectors
}
\note{
This is a very simple function
}
\examples{

myX <- c(3,4,5)
myY <- c(1,2,3)
myDomain <- c(3,5)
integrateIt(myX, myY,myDomain,Trap)
}
\seealso{
\code{\link{print}}
}
\author{
Gangyi Sun
}
