% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Integrating using simpsons rule and trapezoid rule}
\usage{
integrateIt(x, y, ab, rule, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with the same dimensionality as \code{x}.}

\item{ab}{A numeric vector equal to the max and min of \code{x}.}

\item{rule}{A character string, denoting either "Trap" or "Simpson"}
}
\value{
The evaluation of the Simpson or Trapezoid rule, and an object of class Trapezoid or Simpson, depending on \code{rule}, containing
 \item{x}{A numeric vector of x values}
 \item{y}{A numeric vector of corresponding y values, such that y=f(x) is a valid function.} 
 \item{ab}{The range, the max and min of the x values. }
}
\description{
Finds the solution of simpsons/trapezoid rule
}
\note{
This is not a very simple function
}
\examples{

myX <- c(1,2,3,4,5) 
myY <- c(1,2,3,2,1) 
myAB <-c(1,5)
myRule <-"Simpson"
integrateIt(myX, myY, myAB, myRule)
}
\seealso{
\code{\link{printSimp}}
}
\author{
Ian Davis
}
