% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{intregrateIt}
\title{Integral Approxmiation of Values}
\usage{
integrateIt(x = "numeric", y = "numeric", bounds = "numeric",
  rule = "String")
}
\arguments{
\item{x}{A numeric object representing x values}

\item{y}{A numeric object representing y values with the same dimensionality as \code{x}.}

\item{bounds}{a numeric vector of upper and lower bounds to integrate on}

\item{rule}{a String that determines whether to integrate using Simpsons rule or Trapezoid Rule}
}
\value{
An object of either class Trapezoid or Simpson containinig x, y and integral values
\item \code{x} First Input- X values (widths)
\item \code{y} First Input- Y values (heights)
\item \code{approx} Integral Approximation as per Trapezoid Rule
}
\description{
Finds integral approximation using Trapezoid Rule or Simpsons Rules
}
\note{
This is a very simple function
}
\examples{

x = 1:10
y = 11:20
bounds = c(1,10)
rule = "Simpsons"
integrateIt(x,y,bounds,rule)
}
\author{
Jonathan Gross
}
