% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{integrateIt function}
\usage{
integrateIt(x, y, a, b, Rule)
}
\arguments{
\item{x}{This is a vector of x values that the user inputs}

\item{y}{This is a vector of f(x) values that the user inputs}

\item{a}{The starting point from which the user chooses to integrate}

\item{b}{The ending point to which the user chooses to stop the integration}

\item{Rule}{Whether the user is using the Trapezoidal Rule or Simpson's Rule}
}
\value{
An object of class Squares containing
 \item{x}{This is a vector of x values that the user inputs} 
 \item{y}{This is a vector of f(x) values that the user inputs}
 \item{result}{The calculated output when using one of the rules}
}
\description{
integrates a given dataset of values and caluclates the output value based on either the Trapezoidal rule or Simpson's Rule
}
\examples{

}
\author{
Jacob Metz
}

