% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Finding a definite integral under a curve}
\usage{
integrateIt(x, y, a, b, rule = "Trap")
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x} where that is an evaluluation of \code{x} using the function.}

\item{a}{A number indicating the first point in \code{x}}

\item{b}{A number indicating the second point in \code{x}}

\item{rule}{A string taking on values "Trap" and "Simp" indicating which rule should be sued}
}
\value{
An object of class Trapezoid or Simpson containing
 \item{x}{The first object input corresponding to \code{x} in the call}
 \item{y}{The second object input corresonding to \code{y} in the call}
 \item{answer}{The estimate of the integral}
}
\description{
Uses either Simpson's method or Trapezoidal method to approximate an integral
}
\note{
Could optionally put some details about how the two methods estimate integrals here.
}
\examples{
tryx<-seq(-5, 4.99, by=.95)
tryy<-dnorm(tryx)
exampSimp<-integrateIt(tryx, tryy, -5, 5, rule="Simp")
exampTrap<-integrateIt(tryx, tryy, -5, 5, rule="Trap")
}
\author{
Jacob M. Montgomery
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}
}

