% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integral approximation}
\usage{
integrateIt(n, f, ab, rule)
}
\arguments{
\item{n}{Number of segments to estimate the integral. Must be even if using the Simpson method.}

\item{f}{The function you are wanting to estimate the integral of}

\item{ab}{The bounds you are integrating between.}

\item{rule}{Which rule you want to use for estimation; Trapezoid or Simpson.}
}
\value{
A list with the elements
 \item{Object}{The Trapezoid or Simpson integral object}
 \item{coordinates}{The x and y coordinates the integral was estimated using}
 \item{result}{The solution to the integration estimate}
}
\description{
Estimates the integral of a function in specified boundaries
}
\note{
This is a rudimentary implementation of two estimation methods.
}
\examples{

n <- 10
f <- function(x){return(x^2)}
ab <- c(0,10)

integrateIt(n,f,ab,"Trapezoid")
integrateIt(n,f,ab,"Simpson")


}
\author{
Jordon Newton
}
