% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{integrateIt}
\usage{
integrateIt(X, Y, bounds, Rule)
}
\arguments{
\item{X}{A list of even partitions of the bounds that the integral is beign taken on.}

\item{Y}{A numeric list with the same dimensionality as \code{x}, equal to the height of the function at each X value. X and Y form sets of ordered pairs.}

\item{bounds}{The minimum and maximum X values.}

\item{Rule}{Can be either "Trap" or "Simpson". Determines which form of estimation the method will use to estimate the integral}
}
\value{
An object of class Squares containing
 \item{object}{An object of class Simpson or Trapezoid}
}
\description{
integrateIt
}
\note{
This is a very simple function
}
\examples{

myX <- c(1,2,3,4,5)
myY <- c(1,4,9,16,25)
integrateIt(myX, myY, c(1,5), "Trap")
}
\author{
Jonah Klein-Barton
}
