% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt-package.R
\docType{package}
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt-package}
\title{integrateIt}
\description{
The integrateIt package performs approximations of integrats.
}
\examples{

\dontrun{
demo(trapezoid_fun)
demo(simpson_fun)
}

}
\author{
Kimberly Acquilano: \email{k.a.acquilano@wustl.edu} and
}
