% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt-package.R, R/integrateIt.R
\docType{package}
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt-package}
\alias{integrateIt,ANY-method}
\title{integrateIt}
\usage{
integrateIt(Rule, a, b, d, f)
}
\arguments{
\item{Rule}{defining the rule to be used:Trapezoid or Simpson - must use quotes around rule name}

\item{a}{a numeric object of the start value}

\item{b}{a numeric object of the end value}

\item{d}{a numeric object of the number of divisions to be made}

\item{f}{the function}
}
\value{
An object of class Trapezoid or Simpson containing
\item{w}{a vector of values under the integral}
\item{y}{a vector of evaluated values from w}
\item{result}{the approximated result}
}
\description{
The integrateIt package performs approximations of integrals.

Uses the Trapezoid and Simpson Rules to calculate the integral
}
\note{
This produces an object of the Trapezoid or Simpson class
}
\examples{

\dontrun{
demo(trapezoid_fun)
demo(simpson_fun)
}


integrateIt(Rule = "Simpson", a = 0, b = 10, d = 5, f = sin)
}
\seealso{
\code{\link{trapezoid_fun}}, \code{\link{simpson_fun}}
}
\author{
Kimberly Acquilano: \email{k.a.acquilano@wustl.edu} and

Kimberly Acquilano
}
