% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Calculate definite integrals}
\usage{
integrateIt(xvalues, yvalues, start, end, rule)
}
\arguments{
\item{xvalues}{A numeric object}

\item{yvalues}{A numeric object with the same dimensionality as \code{xvalues}.}

\item{start}{A numeric object of length 1}

\item{end}{A numeric object of length 1}

\item{rule}{A character string equal to either "Trapezoid" or "Simpson"}
}
\value{
An object of class Trapezoid or Simpson containing
 \item{xvalues}{The first object input}
 \item{yvalues}{The second object input}
 \item{area}{The area under the curve}
}
\description{
Calculates definite integrals using Simpson and Trapezoid method
}
\note{
The estimate may vary depending on the method of integration chosen
}
\examples{

myX <- c(1:10) 
myY <- c(11:20) 
integrateIT(myX, myY, "Trapezoid")
}
\author{
Kalen J. Davison
}
