% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Compute integral from a to b}
\usage{
integrateIt(X, Y, a, b, Rule)
}
\arguments{
\item{X}{Vector with X values.}

\item{Y}{Vector with f(X) values.}

\item{a}{The starting point}

\item{b}{The ending point}

\item{Rule}{Integration rule}
}
\value{
The output of the function 'integrateIt' is either a Trapezoid or Simpson object with the following elements
\itemize{
\item \code{X} Vector X
\item \code{Y} Vector Y
\item \code{Int} The estimated integral
 }
}
\description{
This function calculates the integral of a function fun between points a and b
using either the Trapezoid or the Simpson method.
}
\details{
The function accepts four arguments: vector X, vector Y, initial point a, final point b, and the rule used to compute the integral.
}
\author{
Miguel Pereira \email{m.pereira@wustl.edu}
}

