% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,}
\alias{ANY-method}
\title{Integration function}
\usage{
integrateIt(xvals = "numeric", yvals = "numeric", a = "numeric",
  b = "numeric", rule = "String")
}
\arguments{
\item{xvals}{A numeric vector}

\item{yvals}{A numeric vector of the same length as \code{xvals}.}

\item{a}{Start index of integration within xvals & yvals}

\item{b}{End index of integration within xvals & yvals}

\item{rule}{A string indicating whether or not to integrate with Trapezoid or Simpson method}
}
\value{
An object of class Trapezoid or Simpson containing
 \item{xvals}{x values from input}
 \item{yvals}{y values from input} 
 \item{integral}{Area under the curve}
}
\description{
Integration function
}
\note{
Area under the curve based off of x&y-coordinates, starting and ending index of integration, and estimator type
}
\examples{

myX <- c(1:10) 
myY <- c(11:20) 
integrateIt(myX, myY, 5, 9, Trapezoid)
}
\seealso{
\code{\link{Trapezoid}} & \code{\link{Simpson}}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
