% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{IntegrateIt}
\title{Integrate via Trapezoid or Simpson's rule}
\usage{
integrateIt(fun, a, b, n, rule)
}
\arguments{
\item{a}{Starting value found in \code{x}}

\item{b}{Ending value found in \code{x}, numeric object}

\item{rule}{Stating estimation rule to use. Either \code{"Trapezoid"} or \code{"Simpson"}}

\item{x}{Vector of X values, numeric}

\item{y}{Vector of Y values in relation to \code{x}, numeric}
}
\value{
A list containing:
\item{EstType}{Estimation Class}
\item{Estimation}{Result of the estimation}
\item{Values}{Matrix of X and Y given in function}
}
\description{
Calculates the integral estimation using either the Trapezoid rule or Simpson's rule
}
\note{
The interval in this is calculated automatically as being 1 - the distance between a and b in the vector \code{x}
}
\examples{

funct <- function(x) {return(sin(x))}
integrateIt(funct, 1, 5, 10, "Trapezoid")
integrateIt(funct, 1, 5, 10, "Simpson")
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}
}
\author{
Peter Bachman
}
