% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{IntegrateIt}
\title{Integrate via Trapezoid or Simpson's rule}
\usage{
integrateIt(x, y, a, b, rule)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
An object of class AllSquares containing
\item{addSquares}{The sum of the squared values}
\item{subtractSquares}{The difference of the squared values}
\item{x}{The first object input}
\item{y}{The second object input}
}
\description{
Calculates the integral estimation using either the Trapezoid rule or Simpson's rule
}
\note{
This produces an object of a new class
}
\examples{

myX <- c(20, 3)
myY <- c(-2, 4.1)
allSquares(myX, myY)
}
\seealso{
\code{\link{Trapezoid}}
}
\author{
Peter Bachman
}
