% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{IntegrateIt}
\title{Integrate via Trapezoid or Simpson's rule}
\usage{
integrateIt(x, y, a, b, rule)
}
\arguments{
\item{x}{Vector of X values, numeric}

\item{y}{Vector of Y values in relation to \code{x}, numeric}

\item{a}{Starting value found in \code{x}}

\item{b}{Ending value found in \code{x}, numeric object}

\item{rule}{Stating estimation rule to use. Either \code{"Trapezoid"} or \code{"Simpson"}}
}
\value{
A list containing:
\item{EstType}{Estimation Class}
\item{Estimation}{Result of the estimation}
\item{Values}{Matrix of X and Y used in estimation}
}
\description{
Calculates the integral estimation using either the Trapezoid rule or Simpson's rule
}
\note{
This produces an object of a new class
}
\examples{

myX <- c(20, 3)
myY <- c(-2, 4.1)
allSquares(myX, myY)
}
\seealso{
\code{\link{Trapezoid}}
}
\author{
Peter Bachman
}
