% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{IntegrateIt}
\title{Integrate via Trapezoid or Simpson's rule}
\usage{
integrateIt(x, y, a, b, rule)
}
\arguments{
\item{x}{Vector of X values, numeric}

\item{y}{Vector of Y values in relation to \code{x}, numeric}

\item{a}{Starting value found in \code{x}}

\item{b}{Ending value found in \code{x}, numeric object}

\item{rule}{Stating estimation rule to use. Either \code{"Trapezoid"} or \code{"Simpson"}}
}
\value{
A list containing:
\item{EstType}{Estimation Class}
\item{Estimation}{Result of the estimation}
\item{Values}{Matrix of X and Y used in estimation}
}
\description{
Calculates the integral estimation using either the Trapezoid rule or Simpson's rule
}
\note{
The interval in this is calculated automatically as being 1 - the distance between a and b in the vector \code{x}
}
\examples{

X <- 1:5
y <- sin(X)
integrateIt(X, Y, 1, 5, "Trapezoid")
integrateIt(X, Y, 2, 4, "Simpson")
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}
}
\author{
Peter Bachman
}
