% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt-package.r, R/integrateIt.R
\docType{package}
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\alias{integrateIt-package}
\title{integrateIt}
\usage{
integrateIt(x, y, a, b, rule, return = T)
}
\arguments{
\item{x}{A numeric vector of x values.}

\item{y}{A numeric vector of y=f(x) values.}

\item{a}{A numeric starting value, corresponding to a value of x, for integration.}

\item{b}{An numeric ending value, corresponding to a value of x, for integration.}

\item{rule}{A character string `Trap' or `Simp' to denote what type of integration approximation to utilize.}
}
\value{
An object of class `Trapezoid' or 'Simpson' containing
\itemize{
\item \code{x} A numeric vector of x values.
\item \code{y} A numeric vector of y=f(x) values.
\item \code{a} A numeric starting value, corresponding to a value of x, for integration.
\item \code{b} An numeric ending value, corresponding to a value of x, for integration.
\item \code{result} The numeric result of integration from a to b of f(x) with respect to x using Simpson's rule.
}
}
\description{
The integrateIt package allows the user to conduct integration using the Trapezoidal or Simpson's approximations.

A user-friendly function to conduct integration using the Trapezoidal rule or Simpson's rule.
}
\examples{
integrateIt(x=1:7, y=2:8, a=1, b=7, rule="Trap")

}
\author{
Andy Stone: \email{arstone@wustl.edu}

Andy Stone: \email{arstone@wustl.edu}
}
\seealso{
\code{\link{Trapezoid}}
}

