% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method,}
\alias{plot,Trap-Simpson-method}
\title{Create an Object with class "Trap" or "Simpson"}
\usage{
integrateIt(x, y, a, b, Trap = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector x}

\item{y}{A numeric vector y (= f(x))}

\item{a}{A starting value (a should be smaller than b)}

\item{b}{An ending value (b should be greater than a)}

\item{Trap}{A rule argument. If TRUE, the Trapezoidal rule is applied. If FALSE, the Simpson rule is applied. Defalt is TRUE.}
}
\value{
An object of class "Trap" or "Simpson" containing
 \item{x}{A vector of x}
 \item{y}{A vector of y}
 \item{result}{The result calculated by the Trapezoidal or Simpson rule}
}
\description{
Create an instance with S4 class "Trap" or "Simpson"
}
\examples{
myx <- 0:10
myy <- cos(3*myx) + sin(2*sin(myx))
mya <- 0
myb <- 10
integrateIt(myx, myy, mya, myb, Trap=TRUE)
integrateIt(myx, myy, mya, myb, Trap=FALSE)

}
\author{
Taishi Muraoka \email{tmuraoka@wustl.edu}
}

