% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integrating a function}
\usage{
integrateIt(x, y, rule, ...)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{rule}{A character string equal to either "Trap" or "Simpson"}
}
\value{
An object of class Squares containing
 \item{x}{The first object input}
 \item{y}{The second object input} 
 \item{area}{The summed area under the curve}
}
\description{
Finds sum of trapezoids or area under parabolas under a curve.
}
\note{
The estimate will depend on which method of integration you choose.
}
\examples{

myX <- c(1, 2) 
myY <- c(2, 4) 
integrateIt(myX, myY, "Trap")
}
\author{
Tyler Brandt
}
