% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{An integral approximation object}
\usage{
integrateIt(x, y, ends, Rule, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of evaluated values (`y[i] = f(x[i])`), with the same length as \code{x}.}

\item{ends}{A numeric vector of length 2, indicating the starting and ending value for integration.}

\item{Rule}{A character string. Either `Trapezoid` or `Simpson` is allowed.}
}
\value{
A list with the elements
 \item{Integrated}{An object of class `Trapezoid` or class `Simpson`}
 \item{input}{x and y} 
 \item{output}{The result of approximation}
}
\description{
Apply the trapezoidal or Simpson's rule to approximate integration
}
\note{
Approximation of integration
}
\examples{
fx <- function(x) {
return(x^3 + x^2 + 1)
}
x <- seq(0,3,0.1)
y <- fx(x)
ends <- c(0,3)
a <- integrateIt(x=x,y=y,ends=ends,Rule="Trapezoid")
integrate(fx,0,3) ## Compared with the result calculated by integrate()
print(a[[1]])
}
\seealso{
print.Trapezoid, print.Simpson
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
