% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integral approximation}
\usage{
integrateIt(x, fun, ends, Rule)
}
\arguments{
\item{x}{A numeric vector that serves as the input for approximation.}

\item{fun}{A function to be integrated.}

\item{ends}{A numeric vector of length 2, indicating the starting and ending value for integration.}

\item{Rule}{A character string. Either `Trapezoid` or `Simpson` is allowed.}
}
\value{
A list with the elements
 \item{Integrated}{An object of class `Trapezoid` or class `Simpson`}
 \item{input}{x and fun} 
 \item{output}{The result of the approximation}
}
\description{
Apply the trapezoidal or Simpson's rule to approximate integration
}
\note{
Approximation of integration
}
\examples{
fx <- function(x) {
return(x^3 + x^2 + 1)
}
x <- seq(0,3,0.1)
ends <- c(0,3)
a <- integrateIt(x=x,fun=fx,ends=ends,Rule="Trapezoid")
integrate(fx,0,3) ## Compared with the result calculated by integrate()
show(a[[1]])
}
\seealso{
tolTest
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
