% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons.R, R/trapezoid.R
\docType{class}
\name{simpsons-class}
\alias{simpsons-class}
\alias{trapezoid-class}
\title{The output of the integrateIt function}
\description{
Object of class simpsons is from the output of the integrateIt function

Object of class trapezoid is from the output of the integrateIt function
}
\details{
An object of class simpsons has the following slots:
\itemize{
\item \code{x} A vector of x values
 \item \code{y} A vector of y values
  \item \code{result} The approximated area under the definite integral
}

An object of class trapezoid has the following slots:
\itemize{
\item \code{x} A vector of x values
 \item \code{y} A vector of y values
  \item \code{result} The approximated area under the definite integral
}
}
\author{
Zoe <\email{zoeang@wustl.edu}>

Zoe <\email{zoeang@wustl.edu}>
}
