% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratePrint.R
\docType{methods}
\name{print,ANY-method}
\alias{print,ANY-method}
\title{Trapazoidal or Simpsons integration result print}
\usage{
\S4method{print}{ANY}(object)
}
\arguments{
\item{object}{An object of either "Trapezoid" or "Simpsons" class}
}
\value{
An object of class Trapezoid or Simpsons, the x and y vectors and the result of the integration
 \item{result}{The original object of class Trapazoid or Simpson}
}
\description{
prints the output of a vector of x and y values integrated via the trapazoidal or simpsons method
}
\note{
This is a very simple print function
}
\examples{

newObject <- integrateIt(x, y, startAndEnd, rule)
print(newObject)
}
\author{
Aaron J. Brezel
}
