% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrapezoid.R
\docType{methods}
\name{plot,Trapezoid-method}
\alias{plot,Trapezoid-method}
\title{Plots objects of Class Trapezoid}
\usage{
\S4method{plot}{Trapezoid}(x, y, ...)
}
\arguments{
\item{x}{object of class Trapezoid}
}
\value{
A graph of the integral that was taken
}
\description{
Inputing an object of class Trapezoid plots an graph of the inregral that was taken
}
\examples{

 t = integrateIt(vector_x = 1:11,
                 vector_y = 1:11,
                 a = 3, b = 7, rule = "Trap")
 plot(t)
}
\author{
David Flaserstein
}

