% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Trapezoid-method}
\alias{print,Trapezoid-method}
\alias{print}
\alias{#'}
\alias{@export}
\alias{setMethod(f="print",}
\alias{signature=}
\alias{"Trapezoid",}
\alias{definition}
\alias{=}
\alias{function(object){}
\alias{return(c("The}
\alias{Trapezoid}
\alias{Integral}
\alias{Approximation}
\alias{Is:}
\alias{",object@approx))}
\alias{}}
\alias{)}
\title{Print}
\usage{
\S4method{print}{Trapezoid}(x)
}
\arguments{
\item{a}{class Trapezoid or Simpson object to print the value of}
}
\value{
The integral approximation of the values as a string
}
\description{
Finds integral approximation using Trapezoid Rule or Simpsons Rules
}
\note{
This is a very simple function
}
\examples{

x = 1:10
y = 11:20
bounds = c(1,10)
rule = "Simpsons"
obj = integrateIt(x,y,bounds,rule)
print(obj)

}
\author{
Jonathan Gross
}
