% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Trapezoid-method}
\alias{print,Trapezoid-method}
\title{print}
\usage{
\S4method{print}{Trapezoid}(x)
}
\arguments{
\item{x}{An object of class Trapezoid or Simpson}
}
\value{
An estimation of the integral
}
\description{
print
}
\note{
This is a very simple function
}
\examples{
myTrapezoid <- new ("Trapezoid", bounds = c(1,5), X = c(1,2,3,4,5), Y = c(1,4,9,16,25), integral = 42)
print(myTrapezoid)
}
\author{
Jonah Klein-Barton
}
