% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Trapezoid-method}
\alias{print,Trapezoid-method}
\title{print}
\usage{
\S4method{print}{Trapezoid}(x)
}
\arguments{
\item{x}{An object of class Trapezoid or Simpson}
}
\value{
An estimation of the integral
}
\description{
print
}
\note{
This is a very simple function
}
\author{
Jonah Klein-Barton
}
