% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print,ANY-method}
\title{Prints the result of integral approximation}
\usage{
print(Rule, a, b, d, f)
}
\arguments{
\item{Rule}{defining the rule to be used:Trapezoid or Simpson - must use quotes around the name of the rule}

\item{a}{a numeric object of the start value}

\item{b}{a numeric object of the end value}

\item{d}{a numeric object of the number of divisions to be made}

\item{f}{the function}
}
\value{
An object of class Trapezoid or Simpson containing
\item{result}{the approximated result}
}
\description{
Prints result from the Trapezoid and Simpson Rules' calculation of the integral
}
\note{
This produces an object of the Trapezoid or Simpson class
}
\examples{

print(Rule = "Simpson", a = 0, b = 10, d = 5, f = sin)
}
\seealso{
\code{\link{trapezoid_fun}}, \code{\link{simpson_fun}}
}
\author{
Kimberly Acquilano
}
