% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print,Simpson-method}
\alias{print,Trapezoid-method}
\title{A method for intgrateIt}
\usage{
print(object)
}
\arguments{
\item{object}{object of class Simpson or Trapezoid}
}
\description{
Method of \code{print} is created by the \code{print} function
}
\author{
Kimberly Acquilano: \email{k.a.acquilano@wustl.edu}
}
