% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Print the Estimation class}
\usage{
print(object)
}
\arguments{
\item{object}{object that is either "Trapezoid" or "Simpson"}
}
\description{
Prints the estimation of either Simpson's or Trapezoidal Rule
}
\examples{
funct <- function(x) {return(sin(x))}
result <- integrateIt(funct, 1, 5, 10, "Trapezoid")
print(result$EstType)

}
