% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print method.R
\docType{methods}
\name{print,Trapezoid-method}
\alias{print,Trapezoid-method}
\alias{addSquares,ANY-method}
\title{Prints}
\usage{
\S4method{print}{Trapezoid}(x)
}
\arguments{
\item{x}{A trap or simpson}
}
\value{
A numeric
}
\description{
Prints
}
\note{
Please Work
}
\examples{

}
\seealso{
\code{\link{subtractSquares}}
}
\author{
Jack Ploshnick
}
