% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_Trapezoid.R
\name{show,Trapezoid-method}
\alias{show,Trapezoid-method}
\title{An integral approximation object}
\usage{
\S4method{show}{Trapezoid}(object)
}
\arguments{
\item{object}{A Trapezoid object}
}
\value{
The integrated value
 \item{Integrated_value}{The integrated value}
}
\description{
Print out the integrated value
}
\note{
Approximation of integration
}
\examples{
fx <- function(x) {
return(x^3 + x^2 + 1)
}
x <- seq(0,3,0.1)
ends <- c(0,3)
a <- integrateIt(x=x,fun=fx,ends=ends,Rule="Trapezoid")
integrate(fx,0,3) ## Compared with the result calculated by integrate()
show(a[[1]])
}
\seealso{
integrateIt
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
