% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpsonClass.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\title{Simpson Rule}
\value{
Simpson Class
}
\description{
A technique for approximating the definite integral.
}
\section{Slots}{

\describe{
\item{\code{fun}}{function defined by the user.}

\item{\code{x}}{vector of numeric values.}

\item{\code{y}}{vector of numeric values.}

\item{\code{result}}{integer value.}
}}

