% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\title{Simpson's Rule}
\value{
Simpson class
}
\description{
Technique for approximating a definite integrals numerically
}
\section{Slots}{

\describe{
\item{\code{x}}{numeric. A vector of values}

\item{\code{f}}{numeric. A vector of evaluated values}

\item{\code{result}}{numeric. Result of the integrated approximation}
}}

\examples{
new("Simpson", x=1:10, f=1:10, result = 49.5)
}
