% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{Simpson}
\title{Simpson integration object}
\description{
objects in this class \code{Simpson} are created with the function \code{integrateIt}
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{integral} The integral
\item \code{x} a vector of values
\item \code{y} a vector of values of same dimensionality as \code{x} 
}
}
\author{
Annie Jarman: \email{a.a.jarman@wustl.edu}
}
