% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\name{check_simpson}
\alias{check_simpson}
\alias{Simpson-initialize,}
\alias{integrateIt-method}
\title{Simpson Integration Object}
\usage{
check_simpson(object)
}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} and \code{print} functions
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} {a vector of values}
\item \code{y} {a vector of evaluated values based on a function with input x}
\item \code{integral} {the numerical approximation of the definite integral}
}
}
\author{
Cecilia Sui: \email{c.sui@wustl.edu}
}
