% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{print}
\title{A Simpson object}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} function.
This object takes in a vector of x and y and estimates the integral between a 
and b using Simpson's rule.
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{vector_x} a vector of x values
\item \code{vector_y} a vector of f(x) values
\item \code{a} starting value of integration
\item \code{b} ending value of integration
\item \code{adjusted_x} vector_x within a and b
\item \code{adjusted_y} vector_y within a and b
\item \code{h} h value used for calculation
\item \code{n} number of values between a and b
\item \code{integrated_value} the integrated value using Simpsons rule
\item \code{scalar} vector of scalars used in calculation
}
}
\examples{
 S = new("Simpson", vector_x = 1:11,
          vector_y = 1:11, a = 3,b = 7)
 s = integrateIt(vector_x = 1:11,
                 vector_y = 1:11,
                 a = 3, b = 7, rule = "Simp")
}
\author{
David P. Flasterstein: \email{davidflasterstein@wustl.edu}
}

