% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\alias{printSimpson}
\title{A squared value object}
\usage{
printSimpson(object = "Simpson")
}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} function.
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{result} The integrated value
\item \code{x} The first input
\item \code{y} the second input 
}
}
\author{
Dominique M. Lockett: \email{dlockett@wustl.edu}
}
