% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{print,Simpson-method}
\title{A Simpson object}
\usage{
\S4method{print}{Simpson}(x, ...)
}
\description{
Object of class \code{Simpson} is created by the \code{integrateIt} function

# getArea method for class Simpson
#' @rdname Simpson
#' @export
setGeneric("getArea",
           function(object="Simpson"){
             standardGeneric("getArea")
             }
)
#' @export
setMethod("getArea", "Simpson",
          function(object){
            return(object@area)
            }
)
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} a vector of x values
\item \code{y} a vector of y values
\item \code{area} area under curve defined by x and y, calculated using the simpsons method
}
}
\author{
Gangyi Sun: \email{gangyi.sun@wustl.edu}
}
