% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{getArea}
\alias{print,Simpson-method}
\title{A Simpson object}
\usage{
getArea(object)

\S4method{print}{Simpson}(x, ...)
}
\description{
Object of class \code{Simpson} is created by the \code{integrateIt} function
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} a vector of x values
\item \code{y} a vector of y values
\item \code{area} area under curve defined by x and y, calculated using the simpsons method
}
}
\author{
Gangyi Sun: \email{gangyi.sun@wustl.edu}
}
