% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\name{validS}
\alias{validS}
\alias{Simpson-class}
\alias{initialize,Simpson}
\alias{validity}
\alias{checker}
\title{A Simpson rule object}
\usage{
validS(object)
}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} functions
}
\author{
Ian Davis
}
