% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize}
\title{A Simpsons Integral Approximation Objejct}
\description{
Object of class \code{Simpson} is created by the \code{integrateIt}
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} First Input- X values (widths)
\item \code{y} First Input- Y values (heights)
\item \code{approx} Integral Approximation as per Simpson Rule
}
}
\author{
Jonathan Gross: \email{jonathan.gross@wustl.edu}
}
