% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simpsons.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{plot,Simpson-method}
\alias{print,Simpson-method}
\title{A Simpson object}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} function
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} The potential values at which the function should be evaluated
\item \code{y} The value of the function evaluated at the corresponding values of x
\item \code{answer} The estimated integral using Simpsons method
}
}
\author{
Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}

