% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{integrateIt,Simpson-method}
\alias{print,Simpson-method}
\title{An approximated value object from Simpson Rule}
\description{
Object of class \code{Smpson} is created by the \code{simpson_fun} function
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{result} The approximated value
\item \code{ab} the start/end values
\item \code{w} the vector of values
\item \code{y} the vector of evaluated values
\item \code{f} the function to be integrated
}
}
\author{
Kimberly Acquilano: \email{k.a.acquilano@wustl.edu}
}
