% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{plot,Simpson-method}
\alias{print,Simpson-method}
\alias{show,Simpson-method}
\title{A Simpson object}
\description{
Objects of class \code{Simpson}
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{X} Vector X
\item \code{Y} Vector f(X)
\item \code{a} initial point to integrate
\item \code{b} final point to integrate 
\item \code{Int} The estimated integral using Simpson's rule 
}
}
\author{
Miguel Pereira: \email{m.pereira@wustl.edu}
}

