% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{plot,Simpson-method}
\alias{print,Simpson-method}
\alias{show,Simpson-method}
\title{A Simpson object}
\description{
Objects of class \code{Candidate} are created by \code{addCandidate} functions
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} Name of the candidate
\item \code{delegatesWon} Number of delegates won by the candidate so far
\item \code{party} Candidate party
\item \code{delegatesNeeded} Number of delegates needed 
}
}
\author{
Miguel Pereira: \email{m.pereira@wustl.edu}
}

