% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{integrateIt,Simpson-method}
\alias{print}
\title{A Simpson class}
\description{
Objects of class \code{Simpson} are created by \code{integrateIt}
}
\details{
\itemize{
\item \code{xvals} a numeric vector
\item \code{yvals} a numeric vector, same dimensions as \code{xvals}
\item \code{integral} area under curve using Trapezoid technique
}
}
\seealso{
\code{\link{integrateIt}}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
