% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{integrateIt,Simpson-method}
\alias{print}
\title{A Simpson class}
\arguments{
\item{xvals}{a numeric vector}

\item{yvals}{a numeric vector, same dimensions as \code{xvals}}

\item{integral}{area under curve using Simpson technique}
}
\description{
Objects of class \code{Simpson} are created by \code{integrateIt}
}
\seealso{
\code{\link{integrateIt}}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
