% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\title{Simpson S4 Class}
\description{
Simpson S4 Class
}
\section{Slots}{

\describe{
\item{\code{x}}{vector of values}

\item{\code{y}}{vector of resulting values}

\item{\code{a}}{Start value of integration}

\item{\code{b}}{End value of integration}

\item{\code{n}}{number of intervals}

\item{\code{est}}{estimate of integral given the above information}
}}

