\name{Simpson}
\alias{Simpson}
\alias{Simpson-class}
\alias{printSimpson}
\alias{printSimpson,Simpson-method}
\title{A Simpson object}
\description{
  Object of class \code{Simpson} are created by the
  \code{initializeIt} function
}
\details{
  An object of the class "Simpson"" has the following
  slots: \itemize{ \item \code{x} The first input
  \item \code{y} the second input \item \code{result} The result }
}
\author{
  Rohan Gupta: \email{rohangupta@wustl.edu}
}