% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{initialize,Simpson-method}
\alias{print,Simpson-method}
\title{A Simpson object}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} function. Objects of
this class have their own \code{print} method.
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} A numeric vector of x values.
\item \code{y} A numeric vector of y=f(x) values.
\item \code{a} A numeric starting value, corresponding to a value of x, for integration.
\item \code{b} A numeric ending value, corresponding to a value of x, for integration.
\item \code{result} The result of integration from a to b of f(x) with respect to x using Simpson's rule.
}
}
\author{
Andy Stone: \email{arstone@wustl.edu}
}

