% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson.R
\docType{class}
\name{simpson-class}
\alias{simpson-class}
\alias{initialize}
\title{A simpson object}
\description{
Object of class \code{simpson} are created by the \code{integrateIt} function
}
\details{
An object of class 'simpson' has the following slots:
\itemize{
\item \code{x} the x values
\item \code{y} the evaluated f(x) values
\item \code{area} estimated area under the curve
}
}
\author{
Tyler Brandt: \email{brandttyler@wustl.edu}
}
