% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_fun.R
\name{simpson_fun}
\alias{simpson_fun}
\alias{simpson_fun,ANY-method}
\title{Numerical Integration Using the Simpson Rule}
\usage{
simpson_fun(a, b, d, f)
}
\arguments{
\item{a}{The start value}

\item{b}{The end value}

\item{d}{The number of divisions}

\item{f}{The function that will be used for integration}
}
\value{
An object of the Simpson class, which has
\item{ab}{vector of starting and ending values}
\item{x}{vector internal values}
\item{y}{vector evaluated values values}
\item{result}{Simpson Rule result}
}
\description{
Numerical Integration Using the Simpson Rule
}
\examples{
simpson_fun(a = 0, b = 10, d = 5, f = sin)
}
\seealso{
\code{\link{trapezoid_fun}}
}
