% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson Class.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{addSquares,ANY-method}
\title{Class of Simpson}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{estimate}{A numeric}
}
\value{
creates Simpson Class
}
\description{
Class of Simpson
}
\note{
Please Work
}
\examples{

}
\seealso{
\code{\link{subtractSquares}}
}
\author{
Jack Ploshnick
}
