% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Improve integration accuracy}
\usage{
tolTest(fx, start_n, tol, interval, rule)
}
\arguments{
\item{fx}{A function defined over \link{a,b}.}

\item{start_n}{The number of desired subintervals over \link{a,b} used to define x values.}

\item{tol}{The tolerated absolute margin of error.}

\item{interval}{A vector containing the lower and upper bounds of integration \link{a,b}.}

\item{rule}{\code{'Simpson'} or \code{'Trapezoid'} method of numerical integration.}
}
\value{
A list object containing all inputs and the final n and associated absolute error
\item{function_input}{The original function}
\item{start_n}{The user-defined tolerated absolute margin of error}
\item{desired_tol}{The user-defined function values over \link{a,b}}
\item{integration_method}{The numerical integration method used}
\item{interval}{The original interval \link{a,b}}
\item{final_n}{Final numer of subintervals that yields a value within the tolerated error margin}
\item{absolute_error}{The final absolute error of the estimare}
}
\description{
Finds appropriate n for integration accuracy.
}
\examples{

test_fx <- function(x){x^2}
tolTest(text_fx, 20, tol=0.001, c(1,3), rule="Trapezoid")

}
\author{
Alma Velazquez
}
