% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Calculates tolerance test}
\usage{
tolTest(fun, a, b, tol = 1e-05, Rule = "Trap", start = 4,
  correct = integrate(fun, a, b))
}
\arguments{
\item{fun}{A that relates x with y.}

\item{a}{The starting point}

\item{b}{The ending point}

\item{tol}{A value for the difference between the approximation and the correct integral accepted.}

\item{Rule}{The approximation rule.}

\item{start}{Initial n}

\item{correct}{correct integral}
}
\value{
A list with the inputs and outputs of the tolerance test.
\itemize{
\item \code{name} Name of the candidate
\item \code{delegatesWon} Number of delegates won by the candidate so far
\item \code{party} Candidate party
\item \code{delegatesNeeded} Number of delegates needed to win the nomination in her party
 }
}
\description{
This function calculates the numbers of sub-divisions (n) necessary to approximate an integral to
correct value.
}
\details{
The function accepts seven arguments: a function, the starting (a) and ending (b) points of the integral,
the approximation rule, the initial n, and the correct integral.
}
\examples{

createCandidate("Sanders", 140, "Dem") 

}
\author{
Miguel Pereira \email{m.pereira@wustl.edu}
}

