% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Calculates tolerance test}
\usage{
tolTest(fun, a, b, tol = 1e-05, Rule = "Trap", start = 4,
  correct = integrate(fun, a, b))
}
\arguments{
\item{name}{A string with the name of the presidential candidate.}

\item{delegatesWon}{A numeric object with the number of delegates already won by the candidate.}

\item{party}{A string with the candidate's party. It accepts either 'Dem' or 'Rep' for the Democratic and Republican parties, respectively.}
}
\value{
An object of class `Candidate' that contains
\itemize{
\item \code{name} Name of the candidate
\item \code{delegatesWon} Number of delegates won by the candidate so far
\item \code{party} Candidate party
\item \code{delegatesNeeded} Number of delegates needed to win the nomination in her party
 }
}
\description{
This function calculates the integral of a function fun between points a and b
using either the Trapezoid or the Simpson method.
}
\details{
The function accepts three arguments: the delegate's name, the number of delegates won and her party.
}
\examples{

createCandidate("Sanders", 140, "Dem") 

}
\author{
Miguel Pereira \email{m.pereira@wustl.edu}
}

