% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Tolerance of integration by the Trapezoidal rule or Simpson's rule.}
\usage{
tolTest(fun, a, b, tolerance, rule, start, correct)
}
\arguments{
\item{fun}{A function f(x).}

\item{tolerance}{Indication of how much error user will allow from true integral value.}

\item{rule}{A character string `Trap' or `Simp' to denote what type of integration approximation to utilize.}

\item{start}{An integer denoting a starting number of intervals.}

\item{correct}{The true answer for the integral.}
}
\value{
An object of class `Trapezoid' or 'Simpson' containing
\itemize{
\item \code{fun} A numeric vector of x values.
\item \code{tolerance} Indication of how much error user will allow from true integral value.
\item \code{rule} A character string `Trap' or `Simp' to denote what type of integration approximation to utilize.
\item \code{start}  An integer denoting a starting number of intervals.
\item \code{correct} The true answer for the integral.
\item \code{finaln} The final n. 
\item \code{abserror} The absolue error of the final estimate.
}
}
\description{
tolTest is a function to conduct integral approximation until reaching a specified level of tolerance.
}
\examples{
easyfun <- function(x){
     x^2
}
tolTest(fun=easyfun, a=1, b=7, start=2, rule="Simp", correct=114, tolerance=1)
}
\author{
Andy Stone: \email{arstone@wustl.edu}
}
\seealso{
\code{\link{integrateIt}}
}

