% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Tolerance Test with class "Tol"}
\usage{
tolTest(fun, tolerance, a, b, Trap = TRUE, start, correct, ...)
}
\arguments{
\item{fun}{A function}

\item{tolerance}{A level of tolerance}

\item{a}{A starting value}

\item{b}{An ending value}

\item{Trap}{A rule argument. If TRUE, the Trapezoidal rule is applied. If FALSE, the Simpson rule is applied. Defalt is TRUE.}

\item{start}{An argument for the number of intervals it should start with. Defalt is 2}

\item{correct}{The correct answer for the integral}
}
\value{
An object of class "Tol" containing
 \item{fun}{A function}
 \item{tolerance}{A level of tolerance}
 \item{a}{A starting value}
 \item{b}{An ending value}
 \item{rule}{A rule argument, Whether the Trapezoidal rule or Simpson rule is used}
 \item{start}{An argument for the number of intervals it should start with}
 \item{correct}{The correct answer for the integral}
 \item{final_n}{The final number of intervals}
 \item{abs_error}{The absolute error of the estimate}
}
\description{
Conduct a tolerance test and return an object with S4 class "Tol"
}
\examples{
myfun <- function(x){
  return(sqrt(sin(x)^3 + 1))
}
mycorrect <- as.numeric(integrate(myfun, 0, 10)$value)
tolTest(myfun, 0.05, 0, 10, Trap=TRUE, 2, mycorrect)
tolTest(myfun, 0.05, 0, 10, Trap=FALSE, 2, mycorrect)

}
\author{
Taishi Muraoka \email{tmuraoka@wustl.edu}
}

