% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\title{Test the Preciseness Required for an Approximation Method}
\usage{
tolTest(
  fun,
  ends,
  tolerance = 0.001,
  Rule = c("Trapezoid", "Simpson"),
  start = 2,
  correct = integrate(fun, ends[1], ends[2])
)
}
\arguments{
\item{fun}{A function to be integrated.}

\item{ends}{A numeric vector of length 2, indicating the starting and ending value for integration.}

\item{tolerance}{A positive number that indicates to what extent the difference between the correct answer and approximation is allowed. The default is 0.001.}

\item{Rule}{A character string. Either `Trapezoid` or `Simpson` is allowed.}

\item{start}{The number of intervals (with equal length) it should start with. The default is 2.}

\item{correct}{The correct answer for the integral. The default is calculated by applying `fun` and `ends` to `integrate()`.}
}
\value{
A list with the elements
 \item{input}{All the specified inputs.}
 \item{n}{The number of intervals required to reach the precision defined by `tolerance`.} 
 \item{abs_error}{The absolute error of the estimate.}
}
\description{
tolTest takes in a function and increase the number of intervals n until the answer it provides
using the specified approximation (i.e., either Trapezoid or Simpson) is within tolerance of the correct answer.
}
\note{
Test the Preciseness Required for an Approximation Method
}
\examples{
fx <- function(x) {
return(x^3 + x^2 + 1)
}
ends <- c(0,3)
test <- tolTest(fun=fx, ends=ends, Rule="Trapezoid")
test 
}
\seealso{
integrateIt
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
