% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapezoidClass.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{Trapezoid Rule}
\description{
A technique for approximating the definite integral.
}
\section{Slots}{

\describe{
\item{\code{fun}}{function defined by the user.}

\item{\code{x}}{vector of numeric values.}

\item{\code{y}}{vector of numeric values.}

\item{\code{result}}{integer value.}
}}

