% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{Trapezoidal Rule}
\value{
Trapezoidal class
}
\description{
Technique for approximating the definite integral
}
\section{Slots}{

\describe{
\item{\code{x}}{numeric. A vector of values}

\item{\code{f}}{numeric. A vector of evaluated values}

\item{\code{result}}{numeric. Result of the integrated approximation}
}}

\examples{
new("Trapezoid", x=1:10, f=1:10, result = 49.5)
}
