% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid-class.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,integrateIt-method}
\alias{print,Trapezoid-method}
\title{An integrated value object using the Trapezoidal Rule}
\usage{
\S4method{print}{Trapezoid}(x)
}
\description{
Objects of class \code{Trapezoid} are created by the \code{integrateIt} function
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{result} The numerical approximation of an integral F(x) over some interval \code{[a,b]} according to the Trapezoidal Rule
\item \code{x} A vector of partition values over the interval \code{[a,b]}
\item \code{y} The function f evaluated over \code{[a,b]}
}
}
\author{
Alma Velazquez: \email{a.m.velazquez@wustl.edu}
}
