% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{Trapezoid}
\title{Trapezoid integration object}
\description{
objects in this class \code{Trapezoid} are created with the function \code{integrateIt}
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{integral} The integral
\item \code{x} a vector of values
\item \code{y} a vector of evaluated values 
}
}
\author{
Annie Jarman: \email{a.a.jarman@wustl.edu}
}
