% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\name{check_trapezoid}
\alias{check_trapezoid}
\alias{Trapezoid-initialize,}
\alias{integrateIt-method}
\title{Trapezoid Integration Object}
\usage{
check_trapezoid(object)
}
\description{
Object of class \code{Trapezoid} are created by the \code{integrateIt} and \code{print} functions
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{x} {a vector of values}
\item \code{y} {a vector of evaluated values}
\item \code{integral} {the numerical approximation of the definite integral}
}
}
\author{
Cecilia Sui: \email{c.sui@wustl.edu}
}
