% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\name{validT}
\alias{validT}
\alias{Trapezoid-class}
\alias{initialize,}
\alias{Trapezoid}
\alias{validity}
\alias{checker}
\title{A Trapezoid rule object}
\usage{
validT(object)
}
\description{
Object of class \code{Trapezoid} are created by the \code{integrateIt} functions
}
\author{
Ian Davis
}
