% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize}
\title{A Trapezoid Integral Approximation Objejct}
\description{
Object of class \code{Trapezoid} is created by the \code{integrateIt}
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{x} First Input- X values (widths)
\item \code{y} First Input- Y values (heights)
\item \code{approx} Integral Approximation as per Trapezoid Rule
}
}
\author{
Jonathan Gross: \email{jonathan.gross@wustl.edu}
}
