% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simpsons.R, R/Trapezoid.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\alias{plot,Trapezoid-method}
\alias{print,Trapezoid-method}
\title{A Simpson object}
\description{
Object of class \code{Simpson} are created by the \code{integrateIt} function

Object of class \code{Trapezoid} are created by the \code{integrateIt} function
}
\details{
An object of the class `Simpson' has the following slots:
\itemize{
\item \code{x} The potential values at which the function should be evaluated
\item \code{y} The value of the function evaluated at the corresponding values of x
\item \code{answer} The estimated integral using Simpsons method
}

An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{x} The potential values at which the function should be evaluated
\item \code{y} The value of the function evaluated at the corresponding values of x
\item \code{answer} The estimated integral using the Trapezoidal method
}
}
\author{
Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}

Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}

