% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{An estimated integral of a function}
\description{
Object of class \code{SquaresPack} are created by the \code{addSquares} and \code{subtractSquares} functions
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{bounds} The lower bound and upper bound of the intergrand
\item \code{X} An ordered list of X values, between a and b
\item \code{Y} An ordered list of Y values, where Yn = F(Xn)
\item \code{integral} An estimate of the integral
}
}
\author{
Jonah Klein-Barton: \email{jonahkleinbarton@gmail.com}
}
