% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,integrateIt-method,print-method}
\alias{integrateIt}
\title{An approximated value object from Trapezoid Rule}
\usage{
integrateIt(object = "Trapezoid")
}
\description{
Object of class \code{Trapezoid} is created by the \code{trapezoid_fun} function
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{result} The approximated value
\item \code{ab} the start/end values
\item \code{w} the vector of values
\item \code{y} the vector of evaluated values
\item \code{f} the function to be integrated
}
}
\author{
Kimberly Acquilano: \email{k.a.acquilano@wustl.edu}
}
