% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\alias{plot,Trapezoid-method}
\alias{print,Trapezoid-method}
\alias{show,Trapezoid-method}
\title{A Trapezoid object}
\description{
Objects of class \code{Trapezoid}
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{X} Vector X
\item \code{Y} Vector f(X)
\item \code{a} initial point to integrate
\item \code{b} final point to integrate 
\item \code{Int} The estimated integral using Trapezoidal's rule 
}
}
\author{
Miguel Pereira: \email{m.pereira@wustl.edu}
}

