% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\alias{integrateIt,Trapezoid-method}
\alias{print}
\title{A Trapezoid class}
\arguments{
\item{xvals}{a numeric vector}

\item{yvals}{a numeric vector, same dimensions as \code{xvals}}

\item{integral}{area under curve using Trapezoid technique}
}
\description{
Objects of class \code{Trapezoid} are created by \code{integrateIt}
}
\seealso{
\code{\link{integrateIt}}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
