% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\alias{integrateIt,Trapezoid-method}
\alias{print}
\title{A Trapezoid class}
\description{
Objects of class \code{Trapezoid} are created by \code{integrateIt}
}
\details{
\itemize{
\item \code{xvals} a numeric vector
\item \code{yvals} a numeric vector, same dimensions as \code{xvals}
\item \code{integral} area under curve using Trapezoid technique
}
}
\seealso{
\code{\link{integrateIt}}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
