% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{Trapezoid S4 Class}
\description{
Trapezoid S4 Class
}
\section{Slots}{

\describe{
\item{\code{x}}{vector of values}

\item{\code{y}}{vector of resulting values}

\item{\code{a}}{Start value of integration}

\item{\code{b}}{End value of integration}
}}

