\name{Trapezoid}
\alias{Trapezoid}
\alias{Trapezoid-class}
\alias{printSimpson}
\alias{printSimpson,Trapezoid-method}
\title{A Trapezoid object}
\description{
  Object of class \code{Trapezoid} are created by the
  \code{initializeIt} function
}
\details{
  An object of the class "Trapezoid"" has the following
  slots: \itemize{ \item \code{x} The first input
  \item \code{y} the second input \item \code{result} The result }
}
\author{
  Rohan Gupta: \email{rohangupta@wustl.edu}
}