% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\title{A Trapezoid object}
\description{
Object of class \code{Trapezoid} are created by the \code{integrateIt} function. Objects of
this class have their own \code{print} method.
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{x} A numeric vector of x values.
\item \code{y} A numeric vector of y=f(x) values.
\item \code{a} A numeric starting value, corresponding to a value of x, for integration.
\item \code{b} A numeric ending value, corresponding to a value of x, for integration.
\item \code{result} The result of integration from a to b of f(x) with respect to x using the Trapezoidal rule.
}
}
\author{
Andy Stone: \email{arstone@wustl.edu}
}

