% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\docType{class}
\name{trapezoid-class}
\alias{trapezoid-class}
\alias{initialize}
\title{A trapezoid object}
\description{
Object of class \code{trapezoid} are created by the \code{integrateIt} function
}
\details{
An object of class 'trapezoid' has the following slots:
\itemize{
\item \code{x} the x values
\item \code{y} the evaluated f(x) values
\item \code{area} estimated area under the curve
}
}
\author{
Tyler Brandt: \email{brandttyler@wustl.edu}
}
